requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
Pillow==10.1.0
stripe==7.4.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
python-decouple==3.8
psycopg2-binary==2.9.7